requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
configparser==5.0.2
credentials==1.1
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dateutil==1.5
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
